% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{bvn}
\alias{bvn}
\title{Bivariate normal dataset with missing.}
\description{
A dataset simulated from "randomized monotone missingness process" and is used to carry out \code{my.em} and \code{my.boot}.
}
