% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{criterion}
\alias{criterion}
\title{Stopping criterion.}
\usage{
criterion(iter, max, tol, exist)
}
\arguments{
\item{iter}{iteration of EM algorithm (numeric)}

\item{max}{maximum iteration of EM algorithm (numeric)}

\item{tol}{convergence tolerance (numeric)}

\item{exist}{existing relative change}
}
\value{
Return \code{TRUE} if the iteration gets the maximum or existing relative change reach the convergence tolerance else return \code{FALSE}.
}
\description{
It is used to decide whether to stop EM algorithm or not.
}
\examples{
\dontrun{
iter <- 10
max <- 100
tol <- 1e-5
eist <- 9.76e-6
criterion()
}
}
