% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{my.boot}
\alias{my.boot}
\title{My bootstrap algorithm.}
\usage{
my.boot(x, max = 100, tol = 1e-05, B)
}
\arguments{
\item{x}{dataset}

\item{max}{maximum iteration of EM algorithm}

\item{tol}{convergence tolerance}

\item{B}{number of bootstrap}
}
\value{
Return the Bootstrap means and the Bootstrap standard errors.
}
\description{
Bootstrap algorithm to compute the variance in EM algorithm
}
\examples{
\dontrun{
data(bvn)
attach(bvn)
B <- 250
my.boot(x=bvn,B=B)
}
}
