% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{my.em}
\alias{my.em}
\title{My EM algorithm.}
\usage{
my.em(x, max = 100, tol = 1e-05, pr)
}
\arguments{
\item{x}{dataset}

\item{max}{maximum iteration of EM algorithm}

\item{tol}{convergence tolerance}

\item{pr}{print the results of each iteration or not}
}
\value{
If \code{pr==TRUE}, print the results of each iteration, return the estimation of theta as \code{theta.em} and the results of each iteration in \code{record}. If \code{pr==FALSE}, just return the estimation of theta as \code{theta.em} and the results of each iteration in \code{record}.
}
\description{
EM algorithm for bivariate normal data
}
\examples{
\dontrun{
data(bvn)
attach(bvn)
pr <- 1
my.em(bvn,max,tol,pr)
}
}
