\name{StatComp-package}
\alias{StatComp-package}
\docType{package}
\title{
  This package provides several functions for EM algorithm for bivariate normal data.
}
\description{
  Function of EM algorithm for bivariate normal data is provided in this package (\code{criterion}, \code{my.em} and \code{my.boot}).
}
\author{
Jiaqi Wu

Maintainer: Jiaqi Wu<wjq138@mail.ustc.edu.cn>
}
\keyword{ package }
